Change checkers format
The format of the hands checker changes.
An hand checker returns the hand in an ordered
list where the first card is the strongest card
of the hand (and so on, as the hand are explained
on any website).
This makes it possible to compare different hands
with the kickers.

None is returned if the hand is not present.